Part3-HBnB-AUTH_&_DB/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
SQLAlchemy==2.0.36
mysqlclient==2.0.3
python-dotenv==0.19.0
Werkzeug==2.2.3
gunicorn==20.1.0
flasgger==0.9.4
mistune==2.0.0
